Move command line argument traits into Device namespace.
